requirements.txt
# LLM & AI
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.6
openai>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.5.0

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright>=1.40.0
selenium>=4.15.0
lxml>=4.9.3

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Data Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# API Framework
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6

# Database & ORM
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
alembic>=1.13.0

# dbt (Data Build Tool)
dbt-core>=1.7.0
dbt-postgres>=1.7.0

# Workflow Orchestration
prefect>=2.14.0

# Vector Database (Optional)
chromadb>=0.4.18

# Environment Management
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
python-dateutil>=2.8.2
tqdm>=4.66.0
click>=8.1.0

# Async Support
aiohttp>=3.9.0
